+++ unisim_lib/unisim/component/cxx/processor/hcs12x/cpu.hh
@@ -153,7 +153,7 @@ public:
 	EBLB(CPU *cpu) { this->cpu = cpu; }
 	~EBLB() { cpu = NULL; }
 
-	inline static string getRegName(uint8_t num) {
+	inline static string getRegName(unsigned int num) {
 		switch (num) {
 		case EBLBRegs::A: return ("A");
 		case EBLBRegs::B: return ("B");
@@ -172,7 +172,7 @@ public:
 		}
 	}
 
-	inline static uint8_t getRegSize(uint8_t num) {
+	inline static unsigned int getRegSize(unsigned int num) {
 		switch (num) {
 		case EBLBRegs::A: return (8);
 		case EBLBRegs::B: return (8);
@@ -197,9 +197,9 @@ public:
 	 * 0x30:TMP1; 0x31:TMP2; 0x32:TMP3;
 	 * 0x04:D; 0x05:X; 0x06:Y; 0x07:SP
 	 */
-	template <class T>	void setter(uint8_t rr, T val); // setter function
-	template <class T>	T getter(uint8_t rr); // getter function
-	template <class T> void exchange(uint8_t rrSrc, uint8_t rrDst);
+	template <class T>	void setter(unsigned int rr, T val); // setter function
+	template <class T>	T getter(unsigned int rr); // getter function
+	template <class T> void exchange(unsigned int rrSrc, unsigned int rrDst);
 
 private:
 	CPU *cpu;
@@ -225,12 +225,12 @@ public:
 	enum REGS_OFFSETS {X, D, Y, SP, PC, A, B, CCRL, CCRH, CCR};
 
 	//#define MAX_INS_SIZE	CodeType::maxsize;
-	static const uint8_t MAX_INS_SIZE = 8;
-	static const uint8_t QUEUE_SIZE = MAX_INS_SIZE;
+	static const unsigned int MAX_INS_SIZE = 8;
+	static const unsigned int QUEUE_SIZE = MAX_INS_SIZE;
 
 
-	inline void queueFlush(uint8_t nByte); // flush is called after prefetch() to advance the queue cursor (first pointer)
-	inline uint8_t* queueFetch(address_t addr, uint8_t* ins, uint8_t nByte);
+	inline void queueFlush(unsigned int nByte); // flush is called after prefetch() to advance the queue cursor (first pointer)
+	inline uint8_t* queueFetch(address_t addr, uint8_t* ins, unsigned int nByte);
 
 private:
 	address_t	queueCurrentAddress;
@@ -238,7 +238,7 @@ private:
 
 	int		queueFirst, queueNElement;
 
-	inline void queueFill(address_t addr, int position, uint8_t nByte);
+	inline void queueFill(address_t addr, int position, unsigned int nByte);
 
 public:
 
@@ -284,7 +284,7 @@ public:
 	//=                    execution handling methods                     =
 	//=====================================================================
 
-	uint8_t step();	// Return the number of cpu cycles consumed by the operation
+	unsigned int step();	// Return the number of cpu cycles consumed by the operation
 	virtual void Sync();
 
 	enum STATES {RUNNING, WAIT, STOP};
@@ -422,8 +422,8 @@ public:
 	inline void    setRegPC(uint16_t val);
 	inline uint16_t getRegPC();
 
-	inline void    setRegTMP(uint8_t index, uint16_t val);
-	inline uint16_t getRegTMP(uint8_t index);
+	inline void    setRegTMP(unsigned int index, uint16_t val);
+	inline uint16_t getRegTMP(unsigned int index);
 
 	bool read(unsigned int offset, const void *buffer, unsigned int data_length);
 	bool write(unsigned int offset, const void *buffer, unsigned int data_length);
@@ -431,7 +431,7 @@ public:
 	/********************************************************************
 	 * *******  Used for Indexed Operations XB: Postbyte Code  **********
 	 * ******************************************************************/
-	inline static string xb_getAddrRegLabel(uint8_t rr) {
+	inline static string xb_getAddrRegLabel(unsigned int rr) {
 		switch (rr) {
 		case 0:
 			return ("X");
@@ -445,7 +445,7 @@ public:
 		return ("unknown");
 	}
 
-	inline static string xb_getAccRegLabel(uint8_t rr) {
+	inline static string xb_getAccRegLabel(unsigned int rr) {
 		switch (rr) {
 		case 0:
 			return ("A");
@@ -457,10 +457,10 @@ public:
 		return ("unknown"); // rr=11 see accumulator D offset indexed-indirect
 	}
 
-	inline uint16_t xb_getAddrRegValue(uint8_t rr);
-	inline void xb_setAddrRegValue(uint8_t rr,uint16_t val);
+	inline uint16_t xb_getAddrRegValue(unsigned int rr);
+	inline void xb_setAddrRegValue(unsigned int rr,uint16_t val);
 
-	inline uint16_t xb_getAccRegValue(uint8_t rr);
+	inline uint16_t xb_getAccRegValue(unsigned int rr);
 	/*************  END  XB  ***************/
 
 	//=====================================================================
@@ -753,7 +753,7 @@ inline void CPU::memWrite16(address_t logicalAddress,uint16_t data, ADDRESS::MOD
 //=                  Registers Acces Routines                          =
 //======================================================================
 
-uint16_t CPU::xb_getAddrRegValue(uint8_t rr) {
+uint16_t CPU::xb_getAddrRegValue(unsigned int rr) {
 	switch (rr) {
 	case 0:
 		return (getRegX());
@@ -767,7 +767,7 @@ uint16_t CPU::xb_getAddrRegValue(uint8_t rr) {
 	throw std::runtime_error("Internal error");
 }
 
-void CPU::xb_setAddrRegValue(uint8_t rr,uint16_t val) {
+void CPU::xb_setAddrRegValue(unsigned int rr,uint16_t val) {
 	switch (rr) {
 	case 0:
 		(setRegX(val)); break;
@@ -783,7 +783,7 @@ void CPU::xb_setAddrRegValue(uint8_t rr,uint16_t val) {
 }
 
 
-uint16_t CPU::xb_getAccRegValue(uint8_t rr) {
+uint16_t CPU::xb_getAccRegValue(unsigned int rr) {
 	switch (rr) {
 	case 0:
 		return (getRegA());
@@ -830,8 +830,8 @@ uint16_t CPU::getRegSP() { return (regSP); }
 void CPU::setRegPC(uint16_t val) { regPC = val; }
 uint16_t CPU::getRegPC() { return (regPC); }
 
-void CPU::setRegTMP(uint8_t index, uint16_t val) { regTMP[index] = val; }
-uint16_t CPU::getRegTMP(uint8_t index) { return (regTMP[index]); }
+void CPU::setRegTMP(unsigned int index, uint16_t val) { regTMP[index] = val; }
+uint16_t CPU::getRegTMP(unsigned int index) { return (regTMP[index]); }
 
 
 } // end of namespace hcs12x
