+++ unisim_lib/unisim/component/cxx/processor/hcs12x/mmc.hh
@@ -96,10 +96,10 @@ public:
 	//=   MEMORY MAP (Logical Memories Offsets) and RESET VALUES OF MMC REGISTERS  =
 	//==============================================================================
 
-	static const uint8_t DIRECT_ADDRESS_SIZE	= 8;	// Number of bits used by the CPU to address DIRECT (max=8)
-	static const uint8_t RAM_ADDRESS_SIZE		= 12;	// Number of bits used by the CPU to address RAM (max=16)
-	static const uint8_t EEPROM_ADDRESS_SIZE	= 10;	// Number of bits used by the CPU to address EEPROM (max=16)
-	static const uint8_t FLASH_ADDRESS_SIZE		= 14;	// Number of bits used by the CPU to address FLASH (max=16)
+	static const unsigned int DIRECT_ADDRESS_SIZE	= 8;	// Number of bits used by the CPU to address DIRECT (max=8)
+	static const unsigned int RAM_ADDRESS_SIZE		= 12;	// Number of bits used by the CPU to address RAM (max=16)
+	static const unsigned int EEPROM_ADDRESS_SIZE	= 10;	// Number of bits used by the CPU to address EEPROM (max=16)
+	static const unsigned int FLASH_ADDRESS_SIZE		= 14;	// Number of bits used by the CPU to address FLASH (max=16)
 
 	static const uint16_t DIRECT_PAGE_SIZE	= 0x100;
 	static const uint16_t RAM_PAGE_SIZE		= 0x1000;
@@ -168,7 +168,7 @@ public:
 
 	enum {MMCCTL0,	MODE, GPAGE, DIRECT, MMCCTL1, RPAGE, EPAGE, PPAGE, RAMWPC, RAMXGU, RAMSHL, RAMSHU};
 
-	static const uint8_t MMC_MEMMAP_SIZE = 12;
+	static const unsigned int MMC_MEMMAP_SIZE = 12;
 	static address_t MMC_REGS_ADDRESSES[MMC_MEMMAP_SIZE];
