+++ unisim_lib/unisim/component/cxx/processor/hcs12x/xgate.hh
@@ -222,7 +222,7 @@ class XGATE :
 
 {
 public:
-	static const uint8_t MAX_INS_SIZE = 2;
+	static const unsigned int MAX_INS_SIZE = 2;
 
 
 	//=========================================================
@@ -288,22 +288,22 @@ public:
 	XGATE(const char *name, Object *parent);
 	~XGATE();
 
-	uint8_t step();	// Return the number of cpu cycles consumed by the operation
+	unsigned int step();	// Return the number of cpu cycles consumed by the operation
 	virtual void Stop(int ret);
 	virtual void Sync();
 
 	virtual void enbale_xgate() = 0;
 	virtual void disable_xgate() = 0;
 
-	virtual void assertInterrupt(uint8_t offset, bool isXGATE_flag = false) = 0;
+	virtual void assertInterrupt(unsigned int offset, bool isXGATE_flag = false) = 0;
 
 	void assert_software_error_interrupt() {
 		assertInterrupt(interrupt_software_error, false);
 	}
 
-	void assertChannelInterrupt(uint8_t offset) {
-		uint8_t reg_index = offset / 8;
-		uint8_t flag_index = offset % 8;
+	void assertChannelInterrupt(unsigned int offset) {
+		unsigned int reg_index = offset / 8;
+		unsigned int flag_index = offset % 8;
 		xgif_register[reg_index] = xgif_register[reg_index] | (1 << flag_index);
 
 		assertInterrupt(offset, true);
@@ -462,10 +462,10 @@ protected:
 	address_t	baseAddress;
 	Parameter<address_t>   param_baseAddress;
 
-	static const uint8_t XGATE_SIZE = 8;
+	static const unsigned int XGATE_SIZE = 8;
 
-	uint8_t sofwtare_channel_id[XGATE_SIZE];
-	ParameterArray<uint8_t> param_software_channel_id;
+	unsigned int sofwtare_channel_id[XGATE_SIZE];
+	ParameterArray<unsigned int> param_software_channel_id;
 
 	bool xgate_enabled;
 	bool stop_current_thread;
@@ -519,9 +519,9 @@ protected:
 		inline void setXGPC(address_t val) { xgpc_register = val; }
 		inline address_t getXGPC() { return (xgpc_register); }
 		inline address_t* getXGPCPtr() { return (&xgpc_register); }
-		inline void setXGRx(uint8_t index, uint16_t val) { xgr_register[index] = ((index == 0)? 0:val); } // R0 is tied to the value 0x0000
-		inline uint16_t getXGRx(uint8_t index) { return (((index == 0)? 0: xgr_register[index])); }
-		inline uint16_t* getXGRxPtr(uint8_t index) { return (&xgr_register[index]); }
+		inline void setXGRx(unsigned int index, uint16_t val) { xgr_register[index] = ((index == 0)? 0:val); } // R0 is tied to the value 0x0000
+		inline uint16_t getXGRx(unsigned int index) { return (((index == 0)? 0: xgr_register[index])); }
+		inline uint16_t* getXGRxPtr(unsigned int index) { return (&xgr_register[index]); }
 		inline void setXGCCR(uint8_t val) { xgccr_register = val & 0x0F; }
 		inline uint8_t getXGCCR() { return (xgccr_register); }
 		inline uint8_t* getXGCCRPtr() { return (&xgccr_register); }
@@ -561,8 +561,8 @@ private:
 	std::map<std::string, Register *> registers_registry;
 	std::vector<unisim::kernel::service::VariableBase*> extended_registers_registry;
 
-	uint8_t				interrupt_software_error;
-	Parameter<uint8_t>	param_interrupt_software_error;
+	unsigned int				interrupt_software_error;
+	Parameter<unsigned int>	param_interrupt_software_error;
 
 	uint64_t	max_inst;
 	Parameter<uint64_t>	   param_max_inst;
@@ -594,7 +594,7 @@ private:
 
 	TSemaphore semphore[8];
 
-	void fetchInstruction(address_t addr, uint8_t* ins, uint8_t nByte);
+	void fetchInstruction(address_t addr, uint8_t* ins, unsigned int nByte);
 
 public:
 
@@ -607,8 +607,8 @@ public:
 	//=                REGISTERS ACCESSORS                    =
 	//=========================================================
 
-	bool lockSemaphore(TOWNER::OWNER owner, uint8_t index);
-	bool unlockSemaphore(TOWNER::OWNER owner, uint8_t index);
+	bool lockSemaphore(TOWNER::OWNER owner, unsigned int index);
+	bool unlockSemaphore(TOWNER::OWNER owner, unsigned int index);
 
 	bool isINTRequestEnabled() { return ((xgmctl_register & 0x0080) != 0); }
 
@@ -628,7 +628,7 @@ public:
 	inline uint16_t getXGISP74() { return (xgvbrPtr_register[2]); }
 	inline void setXGISP74(uint16_t val) { xgvbrPtr_register[2] = val; }
 
-	inline TRegisterBank* getRegisterBank(uint8_t priority) {
+	inline TRegisterBank* getRegisterBank(unsigned int priority) {
 
 		if ((version.compare("V2") != 0) & (priority > 3)) {
 			return (&registerBank[1]);
@@ -637,7 +637,7 @@ public:
 		}
 	}
 
-	void preloadXGR7(uint8_t priority) {
+	void preloadXGR7(unsigned int priority) {
 
 		/**
 		 * XGATE V3 and higher
@@ -659,14 +659,14 @@ public:
 
 	inline void setXGSWT(uint16_t val) { xgswt_register = val; }
 	inline uint16_t getXGSWT() { return (xgswt_register); }
-	inline void setXGRx(uint8_t index, uint16_t val) { currentRegisterBank->setXGRx(index, val); } // R0 is tied to the value 0x0000
-	inline uint16_t getXGRx(uint8_t index) { return (currentRegisterBank->getXGRx(index)); }
+	inline void setXGRx(unsigned int index, uint16_t val) { currentRegisterBank->setXGRx(index, val); } // R0 is tied to the value 0x0000
+	inline uint16_t getXGRx(unsigned int index) { return (currentRegisterBank->getXGRx(index)); }
 	inline void setXGCHID(uint8_t val) { currentRegisterBank->setXGCHID(val); }
 	inline uint8_t getXGCHID() { return (currentRegisterBank->getXGCHID()); }
 	inline void setXGCCR(uint8_t val) { currentRegisterBank->setXGCCR(val); }
 	inline uint8_t getXGCCR() { return (currentRegisterBank->getXGCCR()); }
 
-	inline static std::string getXGRxName(uint8_t index) {
+	inline static std::string getXGRxName(unsigned int index) {
 		stringstream name;
 
 		name << "R" << (unsigned int) index;
