+++ unisim_lib/unisim/component/tlm2/processor/hcs12x/crg.hh
@@ -148,7 +148,7 @@ public:
 	void runCOP();
 	void runClockMonitor();
 
-	void assertInterrupt(uint8_t interrupt_offset);
+	void assertInterrupt(unsigned int interrupt_offset);
 
     //================================================================
     //=                    tlm2 Interface                            =
@@ -259,14 +259,14 @@ private:
 	address_t	baseAddress;
 	Parameter<address_t>   param_baseAddress;
 
-	uint8_t interrupt_offset_rti;
-	Parameter<uint8_t> param_interrupt_offset_rti;
+	unsigned int interrupt_offset_rti;
+	Parameter<unsigned int> param_interrupt_offset_rti;
 
-	uint8_t interrupt_offset_pll_lock;
-	Parameter<uint8_t> param_interrupt_offset_pll_lock;
+	unsigned int interrupt_offset_pll_lock;
+	Parameter<unsigned int> param_interrupt_offset_pll_lock;
 
-	uint8_t interrupt_offset_self_clock_mode;
-	Parameter<uint8_t> param_interrupt_offset_self_clock_mode;
+	unsigned int interrupt_offset_self_clock_mode;
+	Parameter<unsigned int> param_interrupt_offset_self_clock_mode;
 
 	bool	debug_enabled;
 	Parameter<bool>	param_debug_enabled;
